% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIMdynLFD.R
\name{AimDynFLD}
\alias{AimDynFLD}
\title{AIM dynamics with LTBI and deterministic approximation}
\usage{
AimDynFLD(P, tbp, stoch = FALSE, coarse = FALSE, graph = FALSE,
  indat = NULL)
}
\arguments{
\item{P}{}

\item{graph}{}
}
\value{
ans list .... TODO
}
\description{
A function for running the AIM dynamics with infection (fortran version)
}
\details{
Content to be written
}
\author{
Pete Dodd
}

