% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enkf.R
\name{enkf}
\alias{enkf}
\title{Ensemble Kalman-Filter}
\usage{
enkf(X, H, d, eps = 1)
}
\arguments{
\item{X}{}

\item{H}{}

\item{d}{}

\item{eps}{}
}
\value{
X2 a matrix of conditioned data
}
\description{
An Ensemble Kalman Filter
}
\details{
Content to be written
}
\author{
Pete Dodd
}

