% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIMgetCountryParms.R
\name{getCountryAimParms}
\alias{getCountryAimParms}
\title{Parameter preparation for AIM}
\usage{
getCountryAimParms(iso)
}
\arguments{
\item{iso}{}
}
\value{
a list of parameters. See Details.
}
\description{
A function to prepare country specific parameteres for use in demographic/AIM model
}
\details{
Content to be written
}
\author{
Pete Dodd
}

