% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMAP}
\alias{getMAP}
\title{getMAP - optimize to find point (MAP) estimate for model}
\usage{
getMAP(LLP)
}
\arguments{
\item{LLP}{}
}
\value{
list(par,value) with par a list of parameters and value log-posterior
}
\description{
Calibrate model parameters
}
\details{
content to be written
}
\author{
Pete Dodd
}

