% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prepMCMC}
\alias{prepMCMC}
\title{prepMCMC - post-process MCMC data from runMCMC}
\usage{
prepMCMC(X, nsamps = 150, burnin = 150, NW = 20)
}
\arguments{
\item{X}{}

\item{nsamps}{}

\item{burnin}{}

\item{NW}{}
}
\value{
a data.frame of sampled parameters
}
\description{
A function to generate parameter sample from MCMC data
}
\details{
content to be written
}
\author{
Pete Dodd
}

