% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{runMCMC}
\alias{runMCMC}
\title{runMCMC - run MCMC approach to parameter inference}
\usage{
runMCMC(tbp, LLP, NI = 200, NW = 20, previous = NULL)
}
\arguments{
\item{tbp}{}

\item{LLP}{}

\item{NI}{}

\item{NW}{}

\item{previous}{}
}
\value{
data.frame of multiple MCMC chains
}
\description{
Markov-chain Monte Carlo model parameter inference
}
\details{
content to be written
}
\author{
Pete Dodd
}

