% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condsimmer.R
\name{condsimmer}
\alias{condsimmer}
\title{condsimmer - conditional simulation NB needs the same col order in parms as used in ...TODO!}
\usage{
condsimmer(dat = NULL, PL, tbpl, n = 150, quants = c("TBI", "deaths",
  "prev", "prev100k"), obs = c("notes", "dreg", "prev100k"), V = 0,
  show = FALSE, truedat = NULL, fname = NULL, stoch = TRUE,
  coarse = TRUE, west = FALSE)
}
\value{
X a long-form data frame of observed and unobserved quantities
}
\description{
A function for conditional simulation
}
\details{
Content to be written
}
\author{
Pete Dodd
}

